Add new page for android, change xpath, delete some element in HomePage
